requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
albumentations>=1.3.0
tqdm>=4.66.0
Pillow>=10.0.0
ultralytics>=8.1.0
